P31 Phase B checkpoint: ir 5 — lower.py --strings-sites/--strings-slice/--strings-census (p31.tsv consumed as a provenance-headed artifact, fold pcs re-validated), IRExec parses/executes located assignments, cmp, words on the P30 library with lazy literal verification; string_sites.py literal form [@0xW:b, "text"], R[] for indirect EAs; F-B1: EagleString::varying sign-extends the length word (XNLDA) + self-test 1c'; local K=1 gates 0 div (slice 0/1/2/3, k1fo book+stock, k1play book)

+++ Work/c_src/hw/strings/EagleString.cpp
@@ -24,7 +24,12 @@ EagleString EagleString::fixed(uint32_t bp, int32_t n) {
 EagleString EagleString::varying(Memory& memory, uint32_t len_word_addr) {
   EagleString s;
   s.bp = (len_word_addr + 1) * 2;
-  s.len = static_cast<int32_t>(memory.read_word(len_word_addr) & 0xFFFFu);
+  // SIGN-extended: the compiler loads a length word with XNLDA
+  // (EagleGeneral.cpp:51-52, `(src<<16)>>16`), so 0xFFFF is the count -1
+  // (a one-byte DESCENDING string). Found live by P31: DISPLAY_INVENTORY's
+  // compare at 7016816B reads a record field holding 0xFFFF on the login
+  // path; the master runs WCMP with ac1 = -1. (P30 zero-extended.)
+  s.len = static_cast<int32_t>(static_cast<int16_t>(memory.read_word(len_word_addr) & 0xFFFFu));
   return s;
 }
